Updated to use onixlabs-corda-core 2.0.0 and onixlabs-corda-identity-framework 2.0.0 in preparation for v2 features.
