+++ onixlabs-corda-bnms-contract/build.gradle
@@ -28,6 +28,7 @@ dependencies {
 
     // CorDapp Dependencies
     cordapp "$onixlabs_group:onixlabs-corda-core-contract:$onixlabs_corda_core_release_version"
+    cordapp "$onixlabs_group:onixlabs-corda-core-workflow:$onixlabs_corda_core_release_version"
     cordapp "$onixlabs_group:onixlabs-corda-identity-framework-contract:$onixlabs_corda_idfx_release_version"
 
     // Test Dependencies
