+++ onixlabs-corda-bnms-contract/src/main/kotlin/io/onixlabs/corda/bnms/contract/membership/MembershipAttestation.kt
@@ -23,7 +23,7 @@ import io.onixlabs.corda.bnms.contract.membership.MembershipAttestationSchema.Me
 import io.onixlabs.corda.identityframework.contract.Attestation
 import io.onixlabs.corda.identityframework.contract.AttestationPointer
 import io.onixlabs.corda.identityframework.contract.AttestationStatus
-import io.onixlabs.corda.identityframework.contract.toAttestationPointer
+import io.onixlabs.corda.identityframework.contract.toStaticAttestationPointer
 import net.corda.core.contracts.BelongsToContract
 import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.StateRef
@@ -63,7 +63,7 @@ class MembershipAttestation internal constructor(
         membership.state.data.network,
         attestor,
         setOf(membership.state.data.holder),
-        membership.toAttestationPointer(),
+        membership.toStaticAttestationPointer(),
         status,
         metadata,
         linearId,
@@ -103,9 +103,8 @@ class MembershipAttestation internal constructor(
             networkValue = network.value,
             networkOperator = network.operator,
             networkHash = network.hash.toString(),
-            pointerStateRef = pointer.stateRef.toString(),
-            pointerStateClass = pointer.stateClass.canonicalName,
-            pointerStateLinearId = pointer.stateLinearId!!.id,
+            pointer = pointer.statePointer.toString(),
+            pointerStateType = pointer.stateType.canonicalName,
             pointerHash = pointer.hash.toString(),
             status = status,
             previousStateRef = previousStateRef?.toString(),
