+++ onixlabs-corda-bnms-contract/src/main/resources/migration/relationship-attestation-schema.changelog-master.xml
@@ -3,4 +3,5 @@
                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                    xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.5.xsd">
     <include file="migration/relationship-attestation-schema.changelog-v1.xml"/>
+    <include file="migration/relationship-attestation-schema.changelog-v2.xml"/>
 </databaseChangeLog>
