+++ onixlabs-corda-bnms-integration/src/main/kotlin/io/onixlabs/corda/bnms/integration/MembershipAttestationService.kt
@@ -16,14 +16,14 @@
 
 package io.onixlabs.corda.bnms.integration
 
-import io.onixlabs.corda.bnms.contract.Network
-import io.onixlabs.corda.bnms.contract.Setting
 import io.onixlabs.corda.bnms.contract.membership.Membership
-import io.onixlabs.corda.bnms.workflow.membership.AmendMembershipFlow
-import io.onixlabs.corda.bnms.workflow.membership.IssueMembershipFlow
-import io.onixlabs.corda.bnms.workflow.membership.RevokeMembershipFlow
+import io.onixlabs.corda.bnms.contract.membership.MembershipAttestation
+import io.onixlabs.corda.bnms.contract.membership.attest
+import io.onixlabs.corda.bnms.workflow.membership.AmendMembershipAttestationFlow
+import io.onixlabs.corda.bnms.workflow.membership.IssueMembershipAttestationFlow
+import io.onixlabs.corda.bnms.workflow.membership.RevokeMembershipAttestationFlow
 import io.onixlabs.corda.core.integration.RPCService
-import io.onixlabs.corda.identityframework.contract.AbstractClaim
+import io.onixlabs.corda.identityframework.contract.AttestationStatus
 import net.corda.core.contracts.StateAndRef
 import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.identity.AbstractParty
@@ -32,111 +32,111 @@ import net.corda.core.messaging.*
 import net.corda.core.transactions.SignedTransaction
 import java.util.*
 
-class MembershipCommandService(rpc: CordaRPCOps) : RPCService(rpc) {
+class MembershipAttestationService(rpc: CordaRPCOps) : RPCService(rpc) {
 
-    fun issueMembership(
-        network: Network,
-        holder: AbstractParty = ourIdentity,
-        identity: Set<AbstractClaim<*>> = emptySet(),
-        settings: Set<Setting<*>> = emptySet(),
+    fun issueMembershipAttestation(
+        membership: StateAndRef<Membership>,
+        attestor: AbstractParty = ourIdentity,
+        status: AttestationStatus = AttestationStatus.REJECTED,
+        metadata: Map<String, String> = emptyMap(),
         linearId: UniqueIdentifier = UniqueIdentifier(),
         notary: Party? = null,
         observers: Set<Party> = emptySet()
     ): FlowProgressHandle<SignedTransaction> {
-        val membership = Membership(network, holder, identity, settings, linearId)
-        return issueMembership(membership, notary, observers)
+        val attestation = membership.attest(attestor, status, metadata, linearId)
+        return issueMembershipAttestation(attestation, notary, observers)
     }
 
-    fun issueMembership(
-        network: Network,
-        holder: AbstractParty = ourIdentity,
-        identity: Set<AbstractClaim<*>> = emptySet(),
-        settings: Set<Setting<*>> = emptySet(),
+    fun issueMembershipAttestation(
+        membership: StateAndRef<Membership>,
+        attestor: AbstractParty = ourIdentity,
+        status: AttestationStatus = AttestationStatus.REJECTED,
+        metadata: Map<String, String> = emptyMap(),
         linearId: UniqueIdentifier = UniqueIdentifier(),
         notary: Party? = null,
         observers: Set<Party> = emptySet(),
         clientId: String = UUID.randomUUID().toString()
     ): FlowHandleWithClientId<SignedTransaction> {
-        val membership = Membership(network, holder, identity, settings, linearId)
-        return issueMembership(membership, notary, observers, clientId)
+        val attestation = membership.attest(attestor, status, metadata, linearId)
+        return issueMembershipAttestation(attestation, notary, observers, clientId)
     }
 
-    fun issueMembership(
-        membership: Membership,
+    fun issueMembershipAttestation(
+        attestation: MembershipAttestation,
         notary: Party? = null,
         observers: Set<Party> = emptySet()
     ): FlowProgressHandle<SignedTransaction> {
         return rpc.startTrackedFlow(
-            IssueMembershipFlow::Initiator,
-            membership,
+            IssueMembershipAttestationFlow::Initiator,
+            attestation,
             notary,
             observers
         )
     }
 
-    fun issueMembership(
-        membership: Membership,
+    fun issueMembershipAttestation(
+        attestation: MembershipAttestation,
         notary: Party? = null,
         observers: Set<Party> = emptySet(),
         clientId: String = UUID.randomUUID().toString()
     ): FlowHandleWithClientId<SignedTransaction> {
         return rpc.startFlowWithClientId(
             clientId,
-            IssueMembershipFlow::Initiator,
-            membership,
+            IssueMembershipAttestationFlow::Initiator,
+            attestation,
             notary,
             observers
         )
     }
 
-    fun amendMembership(
-        oldMembership: StateAndRef<Membership>,
-        newMembership: Membership,
+    fun amendMembershipAttestation(
+        oldAttestation: StateAndRef<MembershipAttestation>,
+        newAttestation: MembershipAttestation,
         observers: Set<Party> = emptySet()
     ): FlowProgressHandle<SignedTransaction> {
         return rpc.startTrackedFlow(
-            AmendMembershipFlow::Initiator,
-            oldMembership,
-            newMembership,
+            AmendMembershipAttestationFlow::Initiator,
+            oldAttestation,
+            newAttestation,
             observers
         )
     }
 
-    fun amendMembership(
-        oldMembership: StateAndRef<Membership>,
-        newMembership: Membership,
+    fun amendMembershipAttestation(
+        oldAttestation: StateAndRef<MembershipAttestation>,
+        newAttestation: MembershipAttestation,
         observers: Set<Party> = emptySet(),
         clientId: String = UUID.randomUUID().toString()
     ): FlowHandleWithClientId<SignedTransaction> {
         return rpc.startFlowWithClientId(
             clientId,
-            AmendMembershipFlow::Initiator,
-            oldMembership,
-            newMembership,
+            AmendMembershipAttestationFlow::Initiator,
+            oldAttestation,
+            newAttestation,
             observers
         )
     }
 
-    fun revokeMembership(
-        membership: StateAndRef<Membership>,
+    fun revokeMembershipAttestation(
+        attestation: StateAndRef<MembershipAttestation>,
         observers: Set<Party> = emptySet()
     ): FlowProgressHandle<SignedTransaction> {
         return rpc.startTrackedFlow(
-            RevokeMembershipFlow::Initiator,
-            membership,
+            RevokeMembershipAttestationFlow::Initiator,
+            attestation,
             observers
         )
     }
 
-    fun revokeMembership(
-        membership: StateAndRef<Membership>,
+    fun revokeMembershipAttestation(
+        attestation: StateAndRef<MembershipAttestation>,
         observers: Set<Party> = emptySet(),
         clientId: String = UUID.randomUUID().toString()
     ): FlowHandleWithClientId<SignedTransaction> {
         return rpc.startFlowWithClientId(
             clientId,
-            RevokeMembershipFlow::Initiator,
-            membership,
+            RevokeMembershipAttestationFlow::Initiator,
+            attestation,
             observers
         )
     }
